+++ TDEngine2/include/utils/Types.h
@@ -500,10 +500,11 @@ namespace TDEngine2
 		SST_VERTEX,
 		SST_PIXEL,
 		SST_GEOMETRY,
+		SST_COMPUTE,
 		SST_NONE
 	};
 
-	constexpr U8 MaxNumOfShaderStages = static_cast<U8>(SST_GEOMETRY) + 1;
+	constexpr U8 MaxNumOfShaderStages = static_cast<U8>(SST_COMPUTE) + 1;