+++ TDEngine2/plugins/D3D11GraphicsContext/include/CD3D11Shader.h
@@ -109,6 +109,8 @@ namespace TDEngine2
 
 			ID3D11GeometryShader* mpGeometryShader;
 
+			ID3D11ComputeShader*  mpComputeShader;
+
 			TShaderBytecodeDesc   mVertexShaderBytecode;
 	};