+++ TDEngine2/plugins/D3D11GraphicsContext/source/CD3D11Mappings.cpp
@@ -320,6 +320,9 @@ namespace TDEngine2
 			case SST_GEOMETRY:
 				result = "gs_";
 				break;
+			case SST_COMPUTE:
+				result = "cs_";
+				break;
 		}
 
 		switch (version)