+++ TDEngine2/plugins/OGLGraphicsContext/include/COGLShaderCompiler.h
@@ -32,6 +32,8 @@ namespace TDEngine2
 		GLuint mFragmentShaderHandler;
 
 		GLuint mGeometryShaderHandler;
+
+		GLuint mComputeShaderHandler;
 	} TOGLShaderCompilerOutput, *TOGLShaderCompilerOutputPtr;